Issue #4587: Moved resources for translation and arraytypestyle check
