+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/misc/translation/MyLabelsI18.properties
