+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/ButtonLabels.properties
