+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/ButtonLabels_de.translations
