+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/ButtonLabels_es.properties
