+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/ButtonLabels_fr_CA_UNIX.properties
