+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/MyLabelsI18_es.properties
