+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/PageTitles_es.translation
