+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/messages-translation_fr.properties
