+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/messages_test_de.properties
