+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/translation/messages_translation_de.properties
